refactor: ProjectHistoryViewModel 구현
- Controller의 일을 ViewModel로 분리
